- Use rlimits when set. - Fix integer truncation warnings on Solaris.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32705 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/seqdb/seqdbalias.cpp
@@ -329,7 +329,7 @@ void CSeqDBAliasNode::x_ResolveNames(char prot_nucl, CSeqDBLockHold & locked)
         
         for(i = 0; i < m_DBList.size(); i++) {
             CSeqDB_Substring sub(m_DBList[i].GetBasePathS());
-            sub.EraseFront(common + 1);
+            sub.EraseFront((int) common + 1);
             
             m_DBList[i].Assign(sub);
         }
